requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
Pillow==5.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
